requirements.txt
requests==2.31.0
PyYAML==6.0.1
google-cloud-storage==2.14.0
python-dotenv==1.0.0
gunicorn==21.2.0
Flask==2.3.3